Math judge script

+++ nemo_skills/inference/llm_math_judge.py
@@ -22,12 +22,13 @@ import hydra
 from tqdm import tqdm
 
 from nemo_skills.code_execution.sandbox import get_sandbox, sandbox_params
-from nemo_skills.inference.prompt.utils import Prompt, PromptConfig, datasets, prompt_types
 from nemo_skills.inference.server.code_execution_model import (
     ErrorRecoveryConfig,
     get_code_execution_model,
+    get_model,
     server_params,
 )
+from nemo_skills.prompt.utils import get_prompt
 from nemo_skills.utils import get_fields_docstring, get_help_message, nested_dataclass, setup_logging
 
 LOG = logging.getLogger(__file__)
@@ -52,13 +53,12 @@ class GenerateSolutionsConfig:
     server: dict
     # Sandbox configuration {sandbox_params}
     sandbox: dict
-    # Prompt configuration.
-    # Available pre-configured prompts: {prompt_types}.
-    prompt: PromptConfig = field(default_factory=PromptConfig)
+    # Prompt configuration - path to yaml files
+    prompt_template: str | None = None  # not required for OpenAI server
+    prompt_config: str = "judge/math"
     inference: InferenceConfig = field(default_factory=InferenceConfig)  # LLM call parameters
 
     # Can specify one of the existing datasets.
-    # Choices: {datasets}.
     dataset: str | None = None
     split_name: str | None = None  # Can be train, validation, test or train_full (train + validation)
     data_file: str | None = None  # Can directly specify a data file, if using a custom dataset
@@ -71,6 +71,15 @@ class GenerateSolutionsConfig:
     # Useful if need to run multiple slurm jobs on the same data file
     offset: int = 0
 
+    generation_key: str = "judgement"
+
+    # can add this flag to just print the first prompt instead of running generation
+    # useful to double check that your data can be loaded and prompt has what you expect
+    dry_run: bool = False
+
+    # set to True if code execution needs to be supported
+    code_execution: bool = False
+
     def __post_init__(self):
         """Building data_file from dataset/split_name if not provided directly."""
         if self.data_file is not None:
@@ -79,7 +88,16 @@ class GenerateSolutionsConfig:
         else:
             if self.dataset is None or self.split_name is None:
                 raise ValueError("Either `data_file` or `dataset` and `split_name` should be provided")
-            self.data_file = Path(__file__).parents[2] / "datasets" / self.dataset / f"{self.split_name}.jsonl"
+            self.data_file = Path(__file__).parents[1] / "dataset" / self.dataset / f"{self.split_name}.jsonl"
+
+        if self.dataset is None and self.prompt_config is None:
+            raise ValueError("If `dataset` is not provided, `prompt_config` is required")
+
+        if self.server.server_type != "openai" and self.prompt_template is None:
+            raise ValueError("Prompt template is required for non-OpenAI servers")
+
+        if self.server.server_type == "openai" and self.prompt_template is not None:
+            raise ValueError("Prompt template is not supported for OpenAI server")
 
 
 cs = hydra.core.config_store.ConfigStore.instance()
@@ -102,8 +120,11 @@ def generate_solutions(cfg: GenerateSolutionsConfig):
     cfg = GenerateSolutionsConfig(_init_nested=True, **cfg)
 
     LOG.info("Config used: %s", cfg)
-    sandbox = get_sandbox(**cfg.sandbox) if cfg.sandbox is not None else None
-    llm = get_code_execution_model(**cfg.server, sandbox=sandbox)
+    if cfg.code_execution:
+        sandbox = get_sandbox(**cfg.sandbox) if cfg.sandbox is not None else None
+        llm = get_code_execution_model(**cfg.server, sandbox=sandbox)
+    else:
+        llm = get_model(**cfg.server)
 
     # making sure output folder exists
     Path(cfg.output_file).absolute().parent.mkdir(parents=True, exist_ok=True)
@@ -127,10 +148,18 @@ def generate_solutions(cfg: GenerateSolutionsConfig):
 
     # additionally, skipping whatever is pre-filled, assuming offset didn't change
     data = data[starting_idx:]
-    prompt = Prompt(config=cfg.prompt)
+    prompt = get_prompt(cfg.prompt_config, cfg.prompt_template)
+    LOG.info("Prompt used: %s", prompt)
 
     if cfg.max_samples < 0:
         cfg.max_samples = len(data)
+
+    if cfg.prompt_template:
+        LOG.info("Example prompt:\nData dictionary: %s\nPrompt string: %s", data[0], prompt.build_string(data[0]))
+    else:
+        LOG.info("Example prompt:\nData dictionary: %s\nPrompt messages: %s", data[0], prompt.build_messages(data[0]))
+    if cfg.dry_run:
+        return
 
     # setting buffering=1 to force to dump the output after every line, so that we can see intermediate generations
     with open(cfg.output_file, "at" if cfg.skip_filled else "wt", encoding="utf-8", buffering=1) as fout:
@@ -150,13 +179,18 @@ def generate_solutions(cfg: GenerateSolutionsConfig):
                 judgements[-1].update(data_point)
                 prefilled_indices.append(idx)
 
-            if len(data_points) == cfg.batch_size:
-                # batch-computing the outputs
-                outputs = llm.generate(
-                    prompts=[prompt.build_string(data_point) for data_point in data_points],
-                    stop_phrases=list(cfg.prompt.stop_phrases),
-                    **asdict(cfg.inference),
-                )
+            if len(data_points) == cfg.batch_size or idx == cfg.max_samples - 1:
+                if cfg.prompt_template:  # using strings as input if template is provided
+                    outputs = llm.generate(
+                        prompts=[prompt.build_string(data_point) for data_point in data_points],
+                        stop_phrases=list(prompt.config.template.stop_phrases),
+                        **asdict(cfg.inference),
+                    )
+                else:
+                    outputs = llm.generate(
+                        prompts=[prompt.build_messages(data_point) for data_point in data_points],
+                        **asdict(cfg.inference),
+                    )
 
                 prefilled_idx = 0
                 generated_idx = 0
@@ -174,29 +208,6 @@ def generate_solutions(cfg: GenerateSolutionsConfig):
                 judgements = []
                 prefilled_indices = []
 
-        # collecting the final batch
-        if len(data_points) + len(judgements) > 0:
-            if len(data_points) > 0:
-                outputs = llm.generate(
-                    prompts=[prompt.build_string(data_point) for data_point in data_points],
-                    stop_phrases=list(cfg.prompt.stop_phrases),
-                    **asdict(cfg.inference),
-                )
-
-            prefilled_idx = 0
-            generated_idx = 0
-            for cur_idx in range(cfg.max_samples - len(data_points) - len(judgements), cfg.max_samples):
-                if cur_idx in prefilled_indices:
-                    fout.write(json.dumps(judgements[prefilled_idx]) + "\n")
-                    prefilled_idx += 1
-                else:
-                    output, original_data_point = outputs[generated_idx], data_points[generated_idx]
-                    output['judgement'] = output.pop("generation")
-                    original_data_point.pop('judgement', None)
-                    output.update(original_data_point)
-                    fout.write(json.dumps(output) + "\n")
-                    generated_idx += 1
-
 
 error_recovery_params = '\n' + get_fields_docstring(
     ErrorRecoveryConfig,
@@ -207,8 +218,6 @@ error_recovery_params = '\n' + get_fields_docstring(
 
 HELP_MESSAGE = get_help_message(
     GenerateSolutionsConfig,
-    datasets=datasets,
-    prompt_types=prompt_types,
     server_params=server_params(),
     sandbox_params=sandbox_params(),
     error_recovery_params=error_recovery_params,
